requirements.txt
telebot~=0.0.4
requests~=2.28.1
pyTelegramBotAPI~=4.7.0
python-dotenv~=0.21.0